+++ src/ts/components/label.ts
@@ -22,8 +22,11 @@ export interface LabelConfig extends ComponentConfig {
  */
 export class Label<Config extends LabelConfig> extends Component<LabelConfig> {
 
+  private text: string;
+
   private labelEvents = {
-    onClick: new EventDispatcher<Label<Config>, NoArgs>()
+    onClick: new EventDispatcher<Label<Config>, NoArgs>(),
+    onTextChanged: new EventDispatcher<Label<Config>, string>(),
   };
 
   constructor(config: LabelConfig = {}) {
@@ -32,13 +35,15 @@ export class Label<Config extends LabelConfig> extends Component<LabelConfig> {
     this.config = this.mergeConfig(config, {
       cssClass: 'ui-label'
     }, this.config);
+
+    this.text = this.config.text;
   }
 
   protected toDomElement(): DOM {
     let labelElement = new DOM('span', {
       'id': this.config.id,
       'class': this.getCssClasses()
-    }).html(this.config.text);
+    }).html(this.text);
 
     labelElement.on('click', () => {
       this.onClickEvent();
@@ -52,7 +57,17 @@ export class Label<Config extends LabelConfig> extends Component<LabelConfig> {
    * @param text
    */
   setText(text: string) {
+    this.text = text;
     this.getDomElement().html(text);
+    this.onTextChangedEvent(text);
+  }
+
+  /**
+   * Gets the text on this label.
+   * @return {string} The text on the label
+   */
+  getText(): string {
+    return this.text;
   }
 
   /**
@@ -60,10 +75,33 @@ export class Label<Config extends LabelConfig> extends Component<LabelConfig> {
    */
   clearText() {
     this.getDomElement().html('');
+    this.onTextChangedEvent(null);
   }
 
+  /**
+   * Tests if the label is empty and does not contain any text.
+   * @return {boolean} True if the label is empty, else false
+   */
+  isEmpty(): boolean {
+    return !this.text;
+  }
+
+  /**
+   * Fires the {@link #onClick} event.
+   * Can be used by subclasses to listen to this event without subscribing an event listener by overwriting the method
+   * and calling the super method.
+   */
   protected onClickEvent() {
     this.labelEvents.onClick.dispatch(this);
+  }
+
+  /**
+   * Fires the {@link #onClick} event.
+   * Can be used by subclasses to listen to this event without subscribing an event listener by overwriting the method
+   * and calling the super method.
+   */
+  protected onTextChangedEvent(text: string) {
+    this.labelEvents.onTextChanged.dispatch(this, text);
   }
 
   /**
@@ -72,5 +110,13 @@ export class Label<Config extends LabelConfig> extends Component<LabelConfig> {
    */
   get onClick(): Event<Label<LabelConfig>, NoArgs> {
     return this.labelEvents.onClick.getEvent();
+  }
+
+  /**
+   * Gets the event that is fired when the text on the label is changed.
+   * @returns {Event<Label<LabelConfig>, string>}
+   */
+  get onTextChanged(): Event<Label<LabelConfig>, string> {
+    return this.labelEvents.onTextChanged.getEvent();
   }
 }