+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.2.0]
+
+### Added
+- Add `Container#removeComponents()` to remove all child components of a container
+- Display multiple subtitle cues in parallel in `SubtitleOverlay`
+- Add `getText()` method, `isEmpty()` method, and `onTextChanged` event to `Label`
+- Add `TitleBarConfig#keepHiddenWithoutMetadata` to keep `TitleBar` hidden if metadata labels are empty
+
+### Changed
+- Do not display `TitleBar` in Cast UI when it does not contain any metadata (title/description)
+
+### Fixed
+- Clear `SubtitleOverlay` when playback is finished
+
 ## [2.1.1]
 
 ### Fixed
@@ -107,6 +121,7 @@ Version 2.0 of the UI framework is built for player 7.1. If absolutely necessary
 ## 1.0.0 - 2017-02-03
 - First release
 
+[2.2.0]: https://github.com/bitmovin/bitmovin-player-ui/compare/v2.1.1...v2.2.0
 [2.1.1]: https://github.com/bitmovin/bitmovin-player-ui/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/bitmovin/bitmovin-player-ui/compare/v2.0.4...v2.1.0
 [2.0.4]: https://github.com/bitmovin/bitmovin-player-ui/compare/v2.0.3...v2.0.4